backend/app/requirements.txt
fastapi
uvicorn[standard]
motor
reportlab
PyPDF2
requests
python-dotenv
joblib
xgboost
langgraph
pandas
faker
numpy
# scikit-learn==1.7.2
python-multipart
pillow
opencv-python-headless
scikit-image
category-encoders
boto3
langchain
langchain-community
langchain-classic
langchain-core
langchainhub
langchain_groq
langchain_google_genai
langchain-tavily
twilio
# Investor Similarity System - Requirements

# Core Dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
pyyaml>=6.0
copulas>=0.9.0
pgmpy>=0.1.20

# ML Training
xgboost>=2.0.0
# scikit-learn>=1.3.0
shap>=0.44.0

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0

# Type Checking (optional)
mypy>=1.5.0

fastapi
uvicorn[standard]
motor
reportlab
PyPDF2
requests
python-dotenv
joblib
xgboost
langgraph
pandas
faker
numpy
# scikit-learn
joblib
python-multipart
pillow
opencv-python-headless
pandas
scikit-learn
category-encoders
google-genai
joblib
langchain_community
langchain_classic
langchain_groq
langchainhub
langchain-tavily
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
cloudinary==1.44.1
cloudpickle==3.1.2
colorama==0.4.6
copulas==0.14.0
coverage==7.13.1
dataclasses-json==0.6.7
distro==1.9.0
dotenv==0.9.9
fastapi==0.118.0
filelock==3.20.3
frozenlist==1.8.0
fsspec==2026.1.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.3.1
groq==0.37.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
idna==3.10
ImageIO==2.37.2
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-groq==1.1.1
langchain-text-splitters==1.1.0
langchainhub==0.1.21
langsmith==0.6.4
lazy_loader==0.4
librt==0.7.8
llvmlite==0.46.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mpmath==1.3.0
multidict==6.7.0
mypy==1.19.1
mypy_extensions==1.1.0
narwhals==2.7.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.3
opencv-python-headless==4.13.0.90
opt_einsum==3.4.0
orjson==3.11.5
packaging==24.2
pandas==2.3.3
pathspec==1.0.3
patsy==1.0.2
pgmpy==1.0.0
pillow==11.3.0
plotly==6.5.2
pluggy==1.6.0
propcache==0.4.1
protobuf==6.32.1
pyarrow==21.0.0
pydantic==2.11.10
pydantic-settings==2.12.0
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyPDF2==3.0.1
pyro-api==0.1.2
pyro-ppl==1.9.1
pytest==9.0.2
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
reportlab==4.4.4
repoze.lru==0.7
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.27.1
scikit-image==0.26.0
# scikit-learn==1.8.0
scipy==1.17.0
setuptools==80.10.1
shap==0.50.0
six==1.17.0
slicer==0.0.8
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.46
starlette==0.48.0
statsmodels==0.14.6
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2026.1.14
toml==0.10.2
torch==2.10.0
tornado==6.5.2
tqdm==4.67.1
types-requests==2.32.4.20260107
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uuid_utils==0.14.0
uvicorn==0.37.0
watchdog==6.0.0
watchfiles==1.1.1
websockets==16.0
xgboost==3.1.3
yarl==1.22.0
zstandard==0.25.0
langchain_google_genai
langgraph
langchain-classic
langchain-community
langchain-text-splitters
langchain-core
langchain-google-genai
langchain-groq
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
pyyaml>=6.0
copulas>=0.9.0
pgmpy>=0.1.20
xgboost>=2.0.0
# scikit-learn>=1.3.0
shap>=0.44.0
pytest>=7.4.0
pytest-cov>=4.1.0
mypy>=1.5.0
